Moved EventCache and EventFactory into the event-package
