+++ dartagnan/src/main/java/com/dat3m/dartagnan/parsers/program/visitors/boogie/AtomicProcedures.java
@@ -11,8 +11,8 @@ import com.dat3m.dartagnan.program.Register;
 import java.util.Arrays;
 import java.util.List;
 
-import static com.dat3m.dartagnan.program.EventFactory.Atomic;
-import static com.dat3m.dartagnan.program.EventFactory.newStore;
+import static com.dat3m.dartagnan.program.event.EventFactory.Atomic;
+import static com.dat3m.dartagnan.program.event.EventFactory.newStore;
 import static com.dat3m.dartagnan.program.event.lang.catomic.utils.Mo.intToMo;
 
 public class AtomicProcedures {
