+++ dartagnan/src/main/java/com/dat3m/dartagnan/program/Program.java
@@ -4,11 +4,11 @@ package com.dat3m.dartagnan.program;
 import com.dat3m.dartagnan.asserts.AbstractAssert;
 import com.dat3m.dartagnan.configuration.Arch;
 import com.dat3m.dartagnan.program.event.EType;
+import com.dat3m.dartagnan.program.event.EventCache;
 import com.dat3m.dartagnan.program.event.core.Event;
 import com.dat3m.dartagnan.program.filter.FilterBasic;
 import com.dat3m.dartagnan.program.memory.Location;
 import com.dat3m.dartagnan.program.memory.Memory;
-import com.dat3m.dartagnan.program.utils.EventCache;
 import com.google.common.collect.ImmutableSet;
 
 import java.util.ArrayList;
