+++ dartagnan/src/main/java/com/dat3m/dartagnan/program/event/EventFactory.java
@@ -1,10 +1,10 @@
-package com.dat3m.dartagnan.program;
+package com.dat3m.dartagnan.program.event;
 
 import com.dat3m.dartagnan.expression.*;
 import com.dat3m.dartagnan.expression.op.BOpUn;
 import com.dat3m.dartagnan.expression.op.COpBin;
 import com.dat3m.dartagnan.expression.op.IOpBin;
-import com.dat3m.dartagnan.program.event.EType;
+import com.dat3m.dartagnan.program.Register;
 import com.dat3m.dartagnan.program.event.arch.aarch64.StoreExclusive;
 import com.dat3m.dartagnan.program.event.arch.tso.Xchg;
 import com.dat3m.dartagnan.program.event.core.*;
