+++ dartagnan/src/main/java/com/dat3m/dartagnan/program/event/arch/aarch64/StoreExclusive.java
@@ -15,7 +15,7 @@ import com.google.common.base.Preconditions;
 
 import java.util.List;
 
-import static com.dat3m.dartagnan.program.EventFactory.*;
+import static com.dat3m.dartagnan.program.event.EventFactory.*;
 
 public class StoreExclusive extends Store implements RegWriter, RegReaderData {
