+++ dartagnan/src/main/java/com/dat3m/dartagnan/program/event/arch/tso/Xchg.java
@@ -17,7 +17,7 @@ import com.google.common.collect.ImmutableSet;
 
 import java.util.List;
 
-import static com.dat3m.dartagnan.program.EventFactory.*;
+import static com.dat3m.dartagnan.program.event.EventFactory.*;
 
 public class Xchg extends MemEvent implements RegWriter, RegReaderData {
