+++ dartagnan/src/main/java/com/dat3m/dartagnan/program/event/lang/catomic/AtomicFetchOp.java
@@ -13,7 +13,7 @@ import com.dat3m.dartagnan.program.event.core.utils.RegWriter;
 import java.util.List;
 
 import static com.dat3m.dartagnan.configuration.Arch.POWER;
-import static com.dat3m.dartagnan.program.EventFactory.*;
+import static com.dat3m.dartagnan.program.event.EventFactory.*;
 import static com.dat3m.dartagnan.program.event.arch.aarch64.utils.Mo.*;
 import static com.dat3m.dartagnan.program.event.lang.catomic.utils.Mo.SC;
