+++ dartagnan/src/main/java/com/dat3m/dartagnan/program/event/lang/catomic/AtomicThreadFence.java
@@ -6,7 +6,7 @@ import com.dat3m.dartagnan.program.event.core.Fence;
 
 import java.util.List;
 
-import static com.dat3m.dartagnan.program.EventFactory.*;
+import static com.dat3m.dartagnan.program.event.EventFactory.*;
 import static com.dat3m.dartagnan.program.event.lang.catomic.utils.Mo.*;
 
 public class AtomicThreadFence extends Fence {
