+++ dartagnan/src/main/java/com/dat3m/dartagnan/program/event/lang/linux/RMWCmpXchg.java
@@ -17,7 +17,7 @@ import com.google.common.collect.ImmutableSet;
 
 import java.util.List;
 
-import static com.dat3m.dartagnan.program.EventFactory.*;
+import static com.dat3m.dartagnan.program.event.EventFactory.*;
 
 public class RMWCmpXchg extends RMWAbstract implements RegWriter, RegReaderData {
