+++ dartagnan/src/main/java/com/dat3m/dartagnan/program/event/lang/linux/RMWOpAndTest.java
@@ -17,7 +17,7 @@ import com.google.common.base.Preconditions;
 import java.math.BigInteger;
 import java.util.List;
 
-import static com.dat3m.dartagnan.program.EventFactory.*;
+import static com.dat3m.dartagnan.program.event.EventFactory.*;
 
 public class RMWOpAndTest extends RMWAbstract implements RegWriter, RegReaderData {
