+++ dartagnan/src/main/java/com/dat3m/dartagnan/program/event/lang/linux/RMWOpReturn.java
@@ -18,7 +18,7 @@ import com.google.common.base.Preconditions;
 
 import java.util.List;
 
-import static com.dat3m.dartagnan.program.EventFactory.*;
+import static com.dat3m.dartagnan.program.event.EventFactory.*;
 
 public class RMWOpReturn extends RMWAbstract implements RegWriter, RegReaderData {
