+++ dartagnan/src/main/java/com/dat3m/dartagnan/program/event/lang/linux/RMWXchg.java
@@ -16,7 +16,7 @@ import com.google.common.base.Preconditions;
 
 import java.util.List;
 
-import static com.dat3m.dartagnan.program.EventFactory.*;
+import static com.dat3m.dartagnan.program.event.EventFactory.*;
 
 public class RMWXchg extends RMWAbstract implements RegWriter, RegReaderData {
