+++ dartagnan/src/main/java/com/dat3m/dartagnan/program/event/lang/pthread/End.java
@@ -10,7 +10,7 @@ import com.google.common.base.Preconditions;
 
 import java.util.List;
 
-import static com.dat3m.dartagnan.program.EventFactory.*;
+import static com.dat3m.dartagnan.program.event.EventFactory.*;
 import static com.dat3m.dartagnan.program.event.lang.catomic.utils.Mo.SC;
 
 public class End extends Store {
