+++ dartagnan/src/main/java/com/dat3m/dartagnan/program/event/lang/pthread/InitLock.java
@@ -8,8 +8,8 @@ import com.google.common.base.Preconditions;
 
 import java.util.List;
 
-import static com.dat3m.dartagnan.program.EventFactory.eventSequence;
-import static com.dat3m.dartagnan.program.EventFactory.newStore;
+import static com.dat3m.dartagnan.program.event.EventFactory.eventSequence;
+import static com.dat3m.dartagnan.program.event.EventFactory.newStore;
 import static com.dat3m.dartagnan.program.event.lang.catomic.utils.Mo.SC;
 
 public class InitLock extends Store {
