+++ dartagnan/src/main/java/com/dat3m/dartagnan/program/event/lang/pthread/Join.java
@@ -14,8 +14,8 @@ import java.util.ArrayList;
 import java.util.List;
 
 import static com.dat3m.dartagnan.expression.op.COpBin.EQ;
-import static com.dat3m.dartagnan.program.EventFactory.*;
 import static com.dat3m.dartagnan.program.event.EType.PTHREAD;
+import static com.dat3m.dartagnan.program.event.EventFactory.*;
 import static com.dat3m.dartagnan.program.event.lang.catomic.utils.Mo.SC;
 
 public class Join extends Load {
