+++ dartagnan/src/main/java/com/dat3m/dartagnan/program/event/lang/pthread/Unlock.java
@@ -13,9 +13,9 @@ import com.google.common.base.Preconditions;
 import java.util.List;
 
 import static com.dat3m.dartagnan.expression.op.COpBin.NEQ;
-import static com.dat3m.dartagnan.program.EventFactory.*;
 import static com.dat3m.dartagnan.program.event.EType.LOCK;
 import static com.dat3m.dartagnan.program.event.EType.RMW;
+import static com.dat3m.dartagnan.program.event.EventFactory.*;
 import static com.dat3m.dartagnan.program.event.lang.catomic.utils.Mo.SC;
 
 public class Unlock extends MemEvent {
