+++ dartagnan/src/main/java/com/dat3m/dartagnan/program/processing/AtomicAsLock.java
@@ -2,11 +2,11 @@ package com.dat3m.dartagnan.program.processing;
 
 import com.dat3m.dartagnan.expression.Atom;
 import com.dat3m.dartagnan.expression.IConst;
-import com.dat3m.dartagnan.program.EventFactory;
 import com.dat3m.dartagnan.program.Program;
 import com.dat3m.dartagnan.program.Register;
 import com.dat3m.dartagnan.program.Thread;
 import com.dat3m.dartagnan.program.event.EType;
+import com.dat3m.dartagnan.program.event.EventFactory;
 import com.dat3m.dartagnan.program.event.core.Event;
 import com.dat3m.dartagnan.program.event.core.Label;
 import com.dat3m.dartagnan.program.event.lang.svcomp.BeginAtomic;
@@ -18,8 +18,8 @@ import org.sosy_lab.common.configuration.InvalidConfigurationException;
 import java.math.BigInteger;
 
 import static com.dat3m.dartagnan.expression.op.COpBin.NEQ;
-import static com.dat3m.dartagnan.program.EventFactory.newInit;
-import static com.dat3m.dartagnan.program.EventFactory.newLabel;
+import static com.dat3m.dartagnan.program.event.EventFactory.newInit;
+import static com.dat3m.dartagnan.program.event.EventFactory.newLabel;
 
 /**
 Substitutes all atomic markers with accesses to a global mutex.
